Move Ccz4WrappedGr to Ccz4 dir
